updated the major version constant to be an f64 from an f32 to support macos handling floating points for this, untested on other OS's

+++ src-tauri/src/constants.rs
@@ -77,5 +77,5 @@ pub fn get_local_versions_path() -> String {
     format!("{}/local_versions.json", get_config_dir())
 }
 
-pub const MAJOR_VERSION: f32 = 3.3;
+pub const MAJOR_VERSION: f64 = 3.3;
 pub const MINOR_VERSION: i64 = 0;
